requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
# amqp==5.2.0 # Removido - não utilizado
annotated-types==0.6.0
anyio==4.2.0
# appdirs==1.4.4 # Removido - provavelmente transitório ou não essencial
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3 # Usado por y2mate-api indiretamente ou pode ser usado para parsing
# billiard==4.2.0 # Removido - dependência do Celery
Brotli==1.1.0
cachetools==5.3.2
# celery==5.3.6 # Removido - não utilizado
certifi==2023.11.17 # Essencial para HTTPS
# cffi==1.16.0 # Removido - não utilizado diretamente
charset-normalizer==3.3.2 # Dependência do requests
click==8.1.3 # Usado por alguma dependência CLI, manter por segurança ou investigar mais
# click-didyoumean==0.3.1 # Relacionado ao click
# click-plugins==1.1.1 # Relacionado ao click
# click-repl==0.3.0 # Relacionado ao click
colorama==0.4.6 # Usado por alguma dependência CLI
# decorator==4.4.2 # Removido - geral, pode ser transitório
# decouple==0.0.7 # Removido - não parece ser usado para config
# docopt==0.6.2 # Removido - não utilizado
# emoji==2.10.0 # Removido - não utilizado
ffmpeg-progress-yield==0.7.8 # Potencialmente útil se o progresso ffmpeg for desejado
ffmpeg-python==0.2.0 # Usado para interagir com ffmpeg (alternativa ao subprocess direto)
# ffpb==0.4.1 # Removido - alternativa ao ffmpeg-python ou subprocess
frozenlist==1.4.1 # Dependência do aiohttp
# future==0.18.3 # Removido - polyfill Python 2
# google-api-core==2.17.1 # Removido - Google client libs
# google-api-python-client==2.123.0 # Removido
# google-auth==2.28.0 # Removido
# google-auth-httplib2==0.2.0 # Removido
# googleapis-common-protos==1.62.0 # Removido
h11==0.14.0 # Dependência HTTP
httpcore==1.0.2 # Dependência HTTPX
httplib2==0.22.0 # Usado por libs do Google, mas elas foram removidas
httpx==0.26.0 # Cliente HTTP moderno, pode ser usado por dependências
idna==3.6 # Essencial para URLs internacionalizadas
# imageio==2.33.1 # Removido - manipulação de imagem/vídeo, não usado diretamente
# imageio-ffmpeg==0.4.9 # Removido
# kombu==5.3.7 # Removido - dependência do Celery
markdown-it-py==3.0.0 # Usado por pyTelegramBotAPI para parse_mode="Markdown" se usado
mdurl==0.1.2 # Dependência do markdown-it-py
# moviepy==1.0.3 # Removido - edição de vídeo complexa, não usada
multidict==6.0.5 # Dependência do aiohttp
# mutagen==1.47.0 # Removido - metadados de áudio
# numpy==1.26.4 # Removido - científico, não usado diretamente
# outcome==1.3.0.post0 # Removido - controle de fluxo assíncrono, Trio related
# pillow==10.2.0 # Removido - manipulação de imagem, não usado diretamente
# pipreqs==0.4.13 # Removido - ferramenta de geração de requirements
proglog==0.1.10 # Usado por moviepy, mas moviepy removido
# prompt-toolkit==3.0.43 # Removido - CLIs interativas
protobuf==4.25.3 # Usado por libs do Google, mas elas foram removidas
# pyasn1==0.5.1 # Removido - transitório
# pyasn1-modules==0.3.0 # Removido - transitório
# pycparser==2.22 # Removido - parser C, usado por cffi
# pycryptodomex==3.20.0 # Removido - criptografia, não usado diretamente
pydantic==2.6.0 # Usado por y2mate-api
pydantic_core==2.16.1 # Dependência do Pydantic
# pyee==11.1.0 # Removido - pubsub, não usado
Pygments==2.17.2 # Usado por rich ou markdown-it-py
# pygramtic==0.2.0 # Removido
pyparsing==3.1.1 # Dependência comum
PySocks==1.7.1 # Suporte a SOCKS proxy para requests
pyTelegramBotAPI==4.15.4 # Essencial
python-dateutil==2.9.0.post0 # Utilidade de data/hora
python-dotenv==1.0.1 # Essencial
# python-ffmpeg==2.0.10 # Removido - wrapper ffmpeg, já temos ffmpeg-python ou subprocess
# python-telegram-bot==20.8 # Removido - Biblioteca alternativa do Telegram, não usada
# python-telegram-bot-api==0.0.7 # Removido - Outra alternativa de API Telegram
pytube==15.0.0 # Essencial
# redis==5.0.2 # Removido - não utilizado
requests==2.28.2 # Essencial
rich==13.7.0 # Usado por y2mate-api ou outras CLIs
# rq==1.16.1 # Removido - não utilizado
# rsa==4.9 # Removido - criptografia, transitório
# selenium==4.19.0 # Removido - não utilizado
# setuptools==70.0.0 # Essencial para o ambiente Python, mas geralmente não listado explicitamente para apps
six==1.16.0 # Compatibilidade Python 2/3, usado por muitas libs
sniffio==1.3.0 # Usado por anyio/async libs
sortedcontainers==2.4.0 # Estruturas de dados ordenadas
soupsieve==2.5 # Dependência do beautifulsoup4
# tabulate==0.9.0 # Removido - formatação de tabelas, não usado
tqdm==4.65.0 # Barras de progresso, usado por y2mate-api ou pytube
# trio==0.25.0 # Removido - I/O assíncrono, não usado diretamente
# trio-websocket==0.11.1 # Removido
typing_extensions==4.9.0 # Suporte a anotações de tipo
tzdata==2024.1 # Dados de fuso horário para python-dateutil
uritemplate==4.1.1 # Usado por google-api-python-client, mas foi removido
urllib3==1.26.18 # Essencial para requests/aiohttp
# vine==5.1.0 # Removido - dependência do Celery
wcwidth==0.2.13 # Usado por rich/prompt-toolkit
websockets==12.0 # Usado por alguma lib async
wsproto==1.2.0 # Usado por websockets
y2mate-api==1.0.4 # Essencial (y2mate==1.0.0 é uma dependência dela)
# yarg==0.1.9 # Removido - CLI arg parsing
yarl==1.9.4 # Dependência do aiohttp
